requirements.txt
# requirements.txt
# Lists required Python packages and their versions for the project


Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.23